clean up branches/dev-branches routes
